Separate NarrativeLogMenu from SaveMenu
