+++ Assets/Fungus/Scripts/Components/NarrativeLogMenu.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 76befdada4ed8754db75aeb0b0d42976
+timeCreated: 1487446345
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
