+++ Assets/Fungus/Scripts/Signals/WriterSignals.cs
@@ -4,7 +4,7 @@
 namespace Fungus
 {
     /// <summary>
-    /// Writer event signalling system.
+    /// Writer event signaling system.
     /// You can use this to be notified about various events in the writing process.
     /// </summary>
     public static class WriterSignals
